+++ resolwe_bio/tests/processes/test_support_processors.py
@@ -671,4 +671,31 @@ re-save-file report "${NAME}".txt
             alignment_summary,
             'report',
             'hs_gatk_alignment_summary_metrics.txt',
-            file_filter=filter_comment_lines)
+            file_filter=filter_comment_lines
+        )
+
+    @tag_process('insert-size')
+    def test_insert_size(self):
+        with self.preparation_stage():
+            bam = self.run_process('upload-bam', {
+                'src': 'bamclipper/output/TP53.primerclipped.bam',
+                'species': 'Homo sapiens',
+                'build': 'hg19',
+            })
+            genome = self.run_process('upload-genome', {
+                'src': 'bqsr/input/hs_b37_chr17_upto_TP53.fasta.gz',
+                'species': 'Homo sapiens',
+                'build': 'hg19'
+            })
+
+        alignment_summary = self.run_process('insert-size', {
+            'bam': bam.id,
+            'genome': genome.id
+        })
+
+        self.assertFile(
+            alignment_summary,
+            'report',
+            'hs_gatk_insert_size_metrics.txt',
+            file_filter=filter_comment_lines
+        )